Route to receiver based on port rather than appID
I want this to work without requiring both communicating devices to have
the NBSProxy installed. So drop the header communicating the target
APPID. Let apps register for a port, and when on the receiving end we
get data for a port for which an app's registered, launch it to
receive. Apps that are themselves registered to receive will still get
their data, unmodified, even if NBSProxy sent it.

To persist the mapping of port->Set<AppIDs> generalized StatsDB to
store Serializable objects (as Strings).