add callback giving success of register() call
The issue here is that an app on first being installed won't be sent any
Intents until the user's launched it manually. So we'll detect the case
where it seems not to be responding to the REG intent and notify the
client app, which can then put up a notification that if selected
launches the app. Support for that notification is next.