add and check version strings
All IPC Intents now include the version of the sending code. On the app
side, if a client's sending with a version that doesn't match, the
user's prompated to upgrade whichever is older. It's rough, but should
allow a release that doesn't get me in trouble later as things start to
change.