broadcast when app launched or permissions granted
Startup is the problem: user installs the NBSProxy app while a client
app is running. We can prompt the user to launch NBSProxy and then to
grant permissions, but the client app needs to know in order to
immediatly retransmit so the user sees success. So: add to client
library a receiver for messages sent when app launches and then when
permissions are granted.