toward better bootstrap experience
When user first installs NBSProxy the client app will typically be
running and have already called register(). So hold onto the reg data
until able to send it. Try sending again when 1) notified that the app
has launched; and 2) when isInstalled() returns true for the first time.

Also get rid of the WeakReference thing and the requirement that the
Callbacks passed in be instanceof Application. I'm pretty sure that
holding onto a reference in a static won't keep the app from shutting
down and being gc()d. :-)